// src/lib/enums/TextEnums.ts
export enum TEXT_ALIGNMENT {
	Left = 'left',
	Center = 'center',
	Right = 'right',
	Justify = 'justify'
}

export enum HEADER_TAGS {
	H1 = 'h1',
	H2 = 'h2',
	H3 = 'h3',
	H4 = 'h4',
	H5 = 'h5',
	H6 = 'h6'
}

export enum FONT_SIZES {
	Sm = 'sm',
	Md = 'md',
	Lg = 'lg',
	Xl = 'xl',
	'2Xl' = '2xl',
	'3Xl' = '3xl',
	'4Xl' = '4xl',
	'5Xl' = '5xl',
	'6Xl' = '6xl'
}

export enum FONT_WEIGHTS {
	Extralight = 'extralight',
	Light = 'light',
	Medium = 'medium',
	Regular = 'regular',
	Bold = 'bold',
	Extrabold = 'extrabold'
}

export enum TEXT_TRANSFORMS {
	Uppercase = 'uppercase',
	Lowercase = 'lowercase',
	Capitalize = 'capitalize'
}

export enum TEXT_DECORATIONS {
	Underline = 'underline',
	LineThrough = 'line-through',
	NoUnderline = 'no-underline'
}